+++ src/algorithm/array/removeDuplicateLetters.ts
@@ -10,15 +10,16 @@ export function removeDuplicateLetters(s: string): string {
   }
 
   for (let i = 0; i < s.length; i++) {
-    countMap.set(s[i], countMap.get(s[i]) - 1);
+    const item = s[i];
+    countMap.set(item, countMap.get(item) - 1);
 
-    if (!charMap.has(s[i])) {
+    if (!charMap.has(item)) {
       while (stack.length && stack[stack.length - 1].charCodeAt(0) > s.charCodeAt(i) && countMap.get(stack[stack.length - 1]) > 0) {
         charMap.delete(stack.pop());
       }
 
-      charMap.set(s[i], true);
-      stack.push(s[i]);
+      charMap.set(item, true);
+      stack.push(item);
     }
   }