+++ src/algorithm/math/trailingZeroes.ts
@@ -2,10 +2,10 @@
 // 172
 export function trailingZeroes(n: number): number {
   let count = 0;
-  let x = n;
-  while (x) {
-    x = Math.floor(x / 5);
-    count += x;
+  let base = 5;
+  while (base <= n) {
+    count += Math.floor(n / base);
+    base *= 5;
   }
   return count;
 }