+++ src/algorithm/other/majorityElement.ts
@@ -4,8 +4,8 @@ export function majorityElement(nums: number[]): number {
   let index = 0;
   let count = 1;
   for (let i = 1; i < nums.length; i++) {
-    nums[index] === nums[i] ? count++ : count--;
-    if (count === 0) {
+    nums[i] === nums[index] ? count++ : count--;
+    if (count === -1) {
       index = i;
       count = 1;
     }