+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.0.0 (2021-08-17)
+# 0.0.0 (2021-08-22)
 
 ### Bug Fixes
 
@@ -52,6 +52,7 @@
 - **algorithm/array:** add twoSum ([a93da8f](https://github.com/air-supply94/js-algorithm/commit/a93da8fdd1a5cf9228a8cda70410b1e47433fc85))
 - **algorithm/backTrack:** add combine ([74ce4ed](https://github.com/air-supply94/js-algorithm/commit/74ce4eda583e9368d48e1e5948f371ada6064bc8))
 - **algorithm/backTrack:** add findLadders ([ca9ecec](https://github.com/air-supply94/js-algorithm/commit/ca9ecec1ac28ca1b603023f81305c96d02d6f3f3))
+- **algorithm/backTrack:** add letterCombinations ([f31ec42](https://github.com/air-supply94/js-algorithm/commit/f31ec42f4d431315d11a73ee4ce966f0a17c126a))
 - **algorithm/backTrack:** add solveSudoku ([a85601c](https://github.com/air-supply94/js-algorithm/commit/a85601cb2ed535d46e4efe08673a5188250a5f5e))
 - **algorithm/bfs:** add minDepth ([31ead4a](https://github.com/air-supply94/js-algorithm/commit/31ead4a45634cb14bc5ef9151856e93ee790f421))
 - **algorithm/bfs:** add openLock ([59d2a95](https://github.com/air-supply94/js-algorithm/commit/59d2a95acdbb5b2fb1b2f5eb8fceb9faa2582c64))
@@ -86,6 +87,7 @@
 - **array:** add multiplyExceptSelf ([545e5e9](https://github.com/air-supply94/js-algorithm/commit/545e5e9ec432554a9e120f3f40178c8078442b51))
 - **array:** add rotate ([46fd6cf](https://github.com/air-supply94/js-algorithm/commit/46fd6cf9973a42ccf48cac5cd086cc53d6107cec))
 - **array:** add rotate matrix ([cbeb58a](https://github.com/air-supply94/js-algorithm/commit/cbeb58ab6cf44f0f027975ce08e19cab46b1dc98))
+- **array/slidingWindow:** add characterReplacement ([e7410e7](https://github.com/air-supply94/js-algorithm/commit/e7410e78b2d0ba5dc673afa375841094032d8902))
 - **array/slidingWindow:** add maxAverage ([13bd38a](https://github.com/air-supply94/js-algorithm/commit/13bd38a79266365f4d952e7963a83126ffa271df))
 - **avlTree:** 重构此类 ([bd35c86](https://github.com/air-supply94/js-algorithm/commit/bd35c8633f67618017b574312f63328c52a3930b))
 - **AvlTree:** 增加类型 ([57971b0](https://github.com/air-supply94/js-algorithm/commit/57971b0acd7f85c9633708e0c045031c7c415cb7))